chore: Renames storage-pointers to off-chain-data-accessors
