+++ src/off-chain-data-writer.js
@@ -1,0 +1,5 @@
+class OffChainDataWriter {
+
+}
+
+export default OffChainDataWriter;
