+++ test/utils/data-model-definition.js
@@ -1,5 +1,5 @@
 import offChainData from './data/off-chain-data.json';
-import { storageInstance } from '../../src/storage-pointers/in-memory';
+import { storageInstance } from '../../src/off-chain-data-accessors/in-memory';
 
 export const Web3UriBackedDataModel = {
   emptyConfig: {},
